requirements.txt
# langchain
# openai
# streamlit
# python-dotenv
# langchain-community

langchain
langchain-google-genai
streamlit
python-dotenv
